Work around pagination bug in Percipio API.
In the Zoom call today, the Percipio guy said that they changed their API to require pagingRequestId for next pages. Apparently that change also broke pagination if pagingRequestId was empty. So if we only add a pagingRequestId query arg if it's not empty, we get subsequent pages and the connector syncs.